Fix predict method for KFoldsClassifier
Several things here. First, use optimal params from grid search to make
predictions on withheld fold. Second, make folds with stratified k-fold,
which is the same for grid search. Third, track indices and return the
test data and corresponding predicted probabilities in the same order as
the original data, allowing for comparison in the melder.